:art: - Add additional command to run after build finished
This is an issue when running things like a test watcher simultaniously
with Rewatch. Because of the amount of IO Rewacht produces, the watcher
in the secondary thread get's respawned over-and-over, making it produce
more IO than needed. This allows the user to pass in a subcommand to be
ran whenever the build has finished. IE:

```
rewatch watch . -a "yarn jest foo.test"
```

Or, when there are longer build times, alert the user that the build is
finished.

```
rewatch watch . -a "tput bel"
```